fixup! test: add `escapePOSIXShell` util
